requirements.txt
openai>=1.93.1
python-dotenv>=1.0.1
datasets>=2.19.0
tqdm>=4.66.4
pandas>=2.2.2
scikit-learn>=1.3.0
PyYAML>=6.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
